+++ src/pv/format/bufferpercent.c
@@ -24,12 +24,12 @@ size_t pv_formatter_buffer_percent(pvformatter_args_t args)
 		return 0;
 
 	if (args->state->transfer.buffer_size > 0) {
-		int pct_used = pv_percentage((off_t)
-					     (args->state->transfer.read_position -
-					      args->state->transfer.write_position),
-					     (off_t)
-					     (args->state->transfer.buffer_size));
-		(void) pv_snprintf(content, sizeof(content), "{%3d%%}", pct_used);
+		double pct_used = pv_percentage((off_t)
+						(args->state->transfer.read_position -
+						 args->state->transfer.write_position),
+						(off_t)
+						(args->state->transfer.buffer_size));
+		(void) pv_snprintf(content, sizeof(content), "{%3.0f%%}", pct_used);
 	}
 #ifdef HAVE_SPLICE
 	if (args->state->transfer.splice_used)
